requirements.txt
opencv-contrib-python==4.9.0.80
python-telegram-bot==13.15
pyserial==3.5
python-dotenv==1.0.1
numpy==1.26.4

